server/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
python-multipart==0.0.5
scikit-learn==1.0.2
xgboost==1.6.1
librosa==0.9.2
pandas==1.4.2
pytest-shutil==1.7.0
boto3==1.24.40
python-dotenv==0.20.0
google-cloud-storage=2.4.0
